[d3d9] Compress fixed function pixel shader key

+++ src/d3d9/d3d9_fixed_function.h
@@ -112,8 +112,8 @@ namespace dxvk {
       } data;
 
       struct {
-        uint64_t a;
-        uint64_t b;
+        uint32_t a;
+        uint32_t b;
       } primitive;
     };
   };
